requirements.txt
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.0
psycopg2==2.7.7
pycparser==2.19
six==1.12.0
SQLAlchemy<1.4.0
Werkzeug==0.16.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
pytz>=2020.1
Flask-RESTful>=0.3.9
